cnn/requirements.txt
absl-py==0.9.0
astor==0.8.0
certifi==2019.11.28
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.10.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.4
pyparsing==2.4.6
python-dateutil==2.8.1
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
tornado==6.0.3
Werkzeug==1.0.0
wrapt==1.11.2
